requirements.txt
PyQt5>=5.15.0
pycryptodome>=3.20.0
